+++ src/test/java/com/cognite/client/SequencesIntegrationTest.java
@@ -3,11 +3,11 @@ package com.cognite.client;
 import com.cognite.client.config.ClientConfig;
 import com.cognite.client.config.TokenUrl;
 import com.cognite.client.config.UpsertMode;
-import com.cognite.client.dto.Aggregate;
-import com.cognite.client.dto.Item;
-import com.cognite.client.dto.SequenceBody;
-import com.cognite.client.dto.SequenceMetadata;
+import com.cognite.client.dto.*;
 import com.cognite.client.util.DataGenerator;
+import com.google.protobuf.Value;
+import com.google.protobuf.util.Values;
+import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
@@ -17,7 +17,10 @@ import java.time.Duration;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ThreadLocalRandom;
 import java.util.function.BooleanSupplier;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -124,7 +127,7 @@ class SequencesIntegrationTest {
                     Duration.between(startInstant, Instant.now()));
             LOG.info(loggingPrefix + "----------------------------------------------------------------------");
 
-            Thread.sleep(5000); // Wait for evt. consistency
+            Thread.sleep(10000); // Wait for evt. consistency
 
             LOG.info(loggingPrefix + "----------------------------------------------------------------------");
             LOG.info(loggingPrefix + "Start reading sequences headers.");
@@ -182,6 +185,176 @@ class SequencesIntegrationTest {
 
     @Test
     @Tag("remoteCDP")
+    void writeEditReadAndDeleteSequencesRows() throws Exception {
+        Instant startInstant = Instant.now();
+        String loggingPrefix = "UnitTest - writeEditReadAndDeleteSequencesRows() -";
+        LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+        LOG.info(loggingPrefix + "Start test. Creating Cognite client.");
+        CogniteClient client = CogniteClient.ofClientCredentials(
+                        TestConfigProvider.getClientId(),
+                        TestConfigProvider.getClientSecret(),
+                        TokenUrl.generateAzureAdURL(TestConfigProvider.getTenantId()))
+                .withProject(TestConfigProvider.getProject())
+                .withBaseUrl(TestConfigProvider.getHost())
+                //.withClientConfig(config)
+                ;
+        LOG.info(loggingPrefix + "Finished creating the Cognite client. Duration : {}",
+                Duration.between(startInstant, Instant.now()));
+        LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+
+        try {
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+            LOG.info(loggingPrefix + "Start upserting sequences headers.");
+            List<SequenceMetadata> upsertSequencesList = DataGenerator.generateSequenceMetadata(11);
+            client.sequences().upsert(upsertSequencesList);
+            LOG.info(loggingPrefix + "Finished upserting sequences headers. Duration: {}",
+                    Duration.between(startInstant, Instant.now()));
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+
+            Thread.sleep(1000); // Pause... just in case
+
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+            LOG.info(loggingPrefix + "Start upserting sequences rows.");
+            List<SequenceBody> upsertSequenceBodyList = new ArrayList<>();
+            upsertSequencesList.forEach(sequence ->
+                    upsertSequenceBodyList.add(DataGenerator.generateSequenceRows(sequence, 56)));
+            List<SequenceBody> upsertSequenceBodyResponse = client.sequences().rows().upsert(upsertSequenceBodyList);
+            LOG.info(loggingPrefix + "Finished upserting sequences rows. Duration: {}",
+                    Duration.between(startInstant, Instant.now()));
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+
+            Thread.sleep(5000); // Wait for evt. consistency
+
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+            LOG.info(loggingPrefix + "Start reading sequences headers.");
+            List<SequenceMetadata> listSequencesResults = new ArrayList<>();
+            client.sequences()
+                    .list(Request.create()
+                            .withFilterMetadataParameter("source", DataGenerator.sourceValue))
+                    .forEachRemaining(sequences -> listSequencesResults.addAll(sequences));
+
+            BooleanSupplier verifySequenceHeaders = () -> {
+                Map<String, SequenceMetadata> upsertMap = upsertSequencesList.stream()
+                        .collect(Collectors.toMap(SequenceMetadata::getExternalId, Function.identity()));
+                if (upsertMap.size() != listSequencesResults.size()) {
+                    return false;
+                }
+                for (SequenceMetadata sequenceMetadata: listSequencesResults) {
+                    SequenceMetadata original = upsertMap.getOrDefault(sequenceMetadata.getExternalId(), SequenceMetadata.getDefaultInstance());
+                    if (!(sequenceMetadata.getName().equals(original.getName())
+                            && sequenceMetadata.getDescription().equals(original.getDescription())
+                            && sequenceMetadata.getMetadataCount() == original.getMetadataCount()
+                            && sequenceMetadata.getColumnsCount() == original.getColumnsCount()
+                            && sequenceMetadata.getAssetId() == original.getAssetId())) {
+                        return false;
+                    }
+                }
+
+                return true;
+            };
+
+            assertTrue(verifySequenceHeaders, "Sequence header upsert not correct");
+
+            LOG.info(loggingPrefix + "Finished reading sequences headers. Duration: {}",
+                    Duration.between(startInstant, Instant.now()));
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+            LOG.info(loggingPrefix + "Start reading sequences rows.");
+            List<SequenceBody> listSequencesRowsResults = new ArrayList<>();
+            List<Item> sequenceBodyRequestItems = listSequencesResults.stream()
+                    .map(sequenceMetadata -> Item.newBuilder().setId(sequenceMetadata.getId()).build())
+                    .collect(Collectors.toList());
+            client.sequences().rows()
+                    .retrieveComplete(sequenceBodyRequestItems)
+                    .forEachRemaining(sequenceBodies -> listSequencesRowsResults.addAll(sequenceBodies));
+
+            BooleanSupplier verifySequenceRowOriginal = () -> isEqual(upsertSequenceBodyList, listSequencesRowsResults);
+            assertTrue(verifySequenceRowOriginal, "Sequence row upsert not correct");
+
+            LOG.info(loggingPrefix + "Finished reading sequences rows. Duration: {}",
+                    Duration.between(startInstant, Instant.now()));
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+            LOG.info(loggingPrefix + "Start updating sequences rows with new columns.");
+            List<SequenceBody> updatedSequenceBodyList = new ArrayList<>();
+            for (SequenceBody originalBody : upsertSequenceBodyList) {
+                List<SequenceColumn> newColumns = DataGenerator.generateSequenceColumnHeader(2);
+
+                // Add row values for the new columns
+                List<SequenceRow> modifiedRowList = new ArrayList<>();
+                for (SequenceRow row : originalBody.getRowsList()) {
+                    SequenceRow.Builder rowBuilder = row.toBuilder();
+                    for (SequenceColumn column : newColumns) {
+                        if (column.getValueType() == SequenceColumn.ValueType.DOUBLE) {
+                            rowBuilder.addValues(Values.of(ThreadLocalRandom.current().nextDouble(1000000d)));
+                        } else if (column.getValueType() == SequenceColumn.ValueType.LONG) {
+                            rowBuilder.addValues(Values.of(ThreadLocalRandom.current().nextLong(10000000)));
+                        } else {
+                            rowBuilder.addValues(Values.of(RandomStringUtils.randomAlphanumeric(5, 30)));
+                        }
+                    }
+                    modifiedRowList.add(rowBuilder.build());
+                }
+
+                SequenceBody modified = originalBody.toBuilder()
+                        .addAllColumns(newColumns)
+                        .clearRows()
+                        .addAllRows(modifiedRowList)
+                        .build();
+                updatedSequenceBodyList.add(modified);
+            }
+
+            client.sequences().rows().upsert(updatedSequenceBodyList);
+
+            List<SequenceBody> modifiedSequenceBodyResponse = new ArrayList<>();
+            client.sequences().rows()
+                    .retrieveComplete(sequenceBodyRequestItems)
+                    .forEachRemaining(sequenceBodies -> modifiedSequenceBodyResponse.addAll(sequenceBodies));
+
+            BooleanSupplier verifySequenceRowModified = () -> isEqual(updatedSequenceBodyList, modifiedSequenceBodyResponse);
+            assertTrue(verifySequenceRowModified, "Modified sequence row upsert not correct");
+
+            LOG.info(loggingPrefix + "Finished updating sequences rows with new columns. Duration: {}",
+                    Duration.between(startInstant, Instant.now()));
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+
+            Thread.sleep(5000); // Wait for evt. consistency
+
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+            LOG.info(loggingPrefix + "Start deleting sequences rows.");
+            List<SequenceBody> deleteRowsInput = listSequencesRowsResults;
+            List<SequenceBody> deleteRowsResults = client.sequences().rows().delete(deleteRowsInput);
+            LOG.info(loggingPrefix + "Finished deleting sequences rows. Duration: {}",
+                    Duration.between(startInstant, Instant.now()));
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+            LOG.info(loggingPrefix + "Start deleting sequences.");
+            List<Item> deleteItemsInput = new ArrayList<>();
+            listSequencesResults.stream()
+                    .map(sequences -> Item.newBuilder()
+                            .setExternalId(sequences.getExternalId())
+                            .build())
+                    .forEach(item -> deleteItemsInput.add(item));
+
+            List<Item> deleteItemsResults = client.sequences().delete(deleteItemsInput);
+            LOG.info(loggingPrefix + "Finished deleting sequences. Duration: {}",
+                    Duration.between(startInstant, Instant.now()));
+            LOG.info(loggingPrefix + "----------------------------------------------------------------------");
+
+            assertEquals(upsertSequencesList.size(), listSequencesResults.size());
+            assertEquals(deleteItemsInput.size(), deleteItemsResults.size());
+        } catch (Exception e) {
+            LOG.error(e.toString());
+            Thread.sleep(1000);
+            throw new RuntimeException(e);
+        }
+    }
+
+    @Test
+    @Tag("remoteCDP")
     void writeEditAndDeleteSequences() throws Exception {
         Instant startInstant = Instant.now();
         ClientConfig config = ClientConfig.create()
@@ -202,7 +375,7 @@ class SequencesIntegrationTest {
 
         try {
             LOG.info(loggingPrefix + "Start upserting sequences.");
-            List<SequenceMetadata> upsertSequencesList = DataGenerator.generateSequenceMetadata(123);
+            List<SequenceMetadata> upsertSequencesList = DataGenerator.generateSequenceMetadata(12);
             List<SequenceMetadata> upsertedTimeseries = client.sequences().upsert(upsertSequencesList);
             LOG.info(loggingPrefix + "Finished upserting sequences. Duration: {}",
                     Duration.between(startInstant, Instant.now()));
@@ -212,10 +385,26 @@ class SequencesIntegrationTest {
             LOG.info(loggingPrefix + "Start updating sequences.");
             List<SequenceMetadata> editedSequencesInput = upsertedTimeseries.stream()
                     .map(sequences -> sequences.toBuilder()
-                            .setDescription("new-value")
-                            .clearMetadata()
-                            .putMetadata("new-key", "new-value")
-                            .build())
+                                .setDescription("new-value")
+                                .clearMetadata()
+                                .putMetadata("new-key", "new-value")
+                                .addAllColumns(DataGenerator.generateSequenceColumnHeader(2))
+                                .removeColumns(1)
+                                .build())
+                    .map(sequence -> {
+                        // modify all columns
+                        List<SequenceColumn> modifiedColumns = sequence.getColumnsList().stream()
+                                .map(column -> column.toBuilder()
+                                        .clearMetadata()
+                                        .putMetadata("new-column-key", "new-column-value")
+                                        .build())
+                                .collect(Collectors.toList());
+
+                        return sequence.toBuilder()
+                                .clearColumns()
+                                .addAllColumns(modifiedColumns)
+                                .build();
+                    })
                     .collect(Collectors.toList());
 
             List<SequenceMetadata> sequencesUpdateResults = client.sequences().upsert(editedSequencesInput);
@@ -258,7 +447,10 @@ class SequencesIntegrationTest {
                 for (SequenceMetadata sequences : sequencesUpdateResults)  {
                     if (sequences.getDescription().equals("new-value")
                             && sequences.containsMetadata("new-key")
-                            && sequences.containsMetadata(DataGenerator.sourceKey)) {
+                            && sequences.containsMetadata(DataGenerator.sourceKey)
+                            && sequences.getColumnsList().stream().anyMatch(sequenceColumn -> sequenceColumn.containsMetadata(DataGenerator.sourceKey))
+                            && sequences.getColumnsList().stream().anyMatch(sequenceColumn -> sequenceColumn.containsMetadata("new-column-key"))
+                    ) {
                         // all good
                     } else {
                         return false;
@@ -271,7 +463,10 @@ class SequencesIntegrationTest {
                 for (SequenceMetadata sequences : sequencesReplaceResults)  {
                     if (sequences.getDescription().equals("new-value")
                             && sequences.containsMetadata("new-key")
-                            && !sequences.containsMetadata(DataGenerator.sourceKey)) {
+                            && !sequences.containsMetadata(DataGenerator.sourceKey)
+                            && sequences.getColumnsList().stream().allMatch(sequenceColumn -> !sequenceColumn.containsMetadata(DataGenerator.sourceKey))
+                            && sequences.getColumnsList().stream().allMatch(sequenceColumn -> sequenceColumn.containsMetadata("new-column-key"))
+                    ) {
                         // all good
                     } else {
                         return false;
@@ -421,5 +616,73 @@ class SequencesIntegrationTest {
             LOG.error(e.toString());
             throw new RuntimeException(e);
         }
+    }
+
+    private boolean isEqual(List<SequenceBody> left, List<SequenceBody> right) {
+        String loggingPrefix = "isEqual() - ";
+        Map<String, SequenceBody> leftMap = left.stream()
+                .collect(Collectors.toMap(SequenceBody::getExternalId, Function.identity()));
+        if (leftMap.size() != right.size()) {
+            LOG.warn(loggingPrefix + "SequenceBody list inputs not of equal size. Left: {}, right: {}",
+                    left.size(),
+                    right.size());
+            return false;
+        }
+        for (SequenceBody rightBody: right) {
+            // Check the basic content counts.
+            SequenceBody leftBody = leftMap.getOrDefault(rightBody.getExternalId(), SequenceBody.getDefaultInstance());
+            if (!(rightBody.getColumnsCount() == leftBody.getColumnsCount()
+                    && rightBody.getRowsCount() == leftBody.getRowsCount())) {
+                LOG.warn(loggingPrefix + "SequenceBody column and row count not of equal size. Left column: {}, "
+                        + "right column: {}, left row: {}, right row: {}",
+                        leftBody.getColumnsCount(),
+                        rightBody.getColumnsCount(),
+                        leftBody.getRowsCount(),
+                        rightBody.getRowsCount());
+                return false;
+            }
+
+            // Check the column schema.
+            List<SequenceColumn> rightColumns = rightBody.getColumnsList();
+            Map<String, SequenceColumn> leftColumns = leftBody.getColumnsList().stream()
+                    .collect(Collectors.toMap(SequenceColumn::getExternalId, Function.identity()));
+            for (SequenceColumn rightColumn : rightColumns) {
+                if (!leftColumns.containsKey(rightColumn.getExternalId())) {
+                    LOG.warn(loggingPrefix + "SequenceBody column schema not equal. Left side does not have column extId: {}" + System.lineSeparator()
+                                    + "right column: {}",
+                            rightColumn.getExternalId(),
+                            rightColumn);
+                    return false;
+                }
+            }
+
+            // Check the row values.
+            List<SequenceRow> rightRows = rightBody.getRowsList();
+            Map<Long, SequenceRow> leftRows = leftBody.getRowsList().stream()
+                    .collect(Collectors.toMap(SequenceRow::getRowNumber, Function.identity()));
+            for (SequenceRow row : rightRows) {
+                if (!(leftRows.containsKey(row.getRowNumber())
+                        && leftRows.get(row.getRowNumber()).getValuesCount() == row.getValuesCount())) {
+                    LOG.warn(loggingPrefix + "SequenceBody rows counts not equal for row no {}. Left value count: {}, "
+                                    + "right value count: {}",
+                            row.getRowNumber(),
+                            leftRows.get(row.getRowNumber()).getValuesCount(),
+                            row.getValuesCount());
+                    return false;
+                }
+                for (int i = 0; i < row.getValuesCount(); i++) {
+                    if (!row.getValues(i).equals(leftRows.get(row.getRowNumber()).getValues(i))) {
+                        LOG.warn(loggingPrefix + "SequenceBody column value not equal for index {}. Left column: {}" + System.lineSeparator()
+                                        + "right column: {}",
+                                i,
+                                leftRows.get(row.getRowNumber()).getValues(i),
+                                row.getValues(i));
+                        return false;
+                    }
+                }
+            }
+        }
+
+        return true;
     }
 }